+++ Plugin.cs
@@ -11,7 +11,7 @@ public sealed class Plugin : BaseUnityPlugin
 {
     public const string PluginGuid = "com.xman0922.unifiedtargetportal";
     public const string PluginName = "Unified Target Portal";
-    public const string PluginVersion = "1.0.0";
+    public const string PluginVersion = "1.0.1";
 
     internal static ManualLogSource Log = null!;